refactor: densify nine more app-test allowlist suites under 400 lines
Extract shared fixtures and concern-split sibling suites so nine more
app/test files drop to the 400-line cap without new mechanical partN shards.
