+++ app/test/diplomacy_panel_orders_test.dart
@@ -11,29 +11,6 @@ import 'diplomacy_panel_orders_pump_support.dart';
 import 'panel_test_fixtures.dart';
 import 'widget_test_assets.dart';
 
-Game _colonyGame({List<BoycottState> boycotts = const []}) {
-  return buildDiplomacyPanelTestGame().copyWith(
-    colonyStates: const [
-      ColonyState(tribeId: 't1', colonyOfGpId: diplomacyOrdersHumanId, sinceTurn: 1),
-    ],
-    tribes: const [Tribe(id: 't1', displayName: 'Aztec')],
-    boycottStates: boycotts,
-  );
-}
-
-Game _minorEmbassyGame({List<SubsidyState> subsidies = const []}) {
-  return buildDiplomacyRichPanelTestGame().copyWith(
-    overtureStates: const [
-      OvertureState(
-        gpId: diplomacyOrdersHumanId,
-        targetId: diplomacyOrdersMinorId,
-        stage: OvertureStage.embassy,
-      ),
-    ],
-    subsidyStates: subsidies,
-  );
-}
-
 void main() {
   suppressLogsForTests();
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -72,12 +49,16 @@ void main() {
       final game = buildDiplomacyPanelTestGame();
       final humanId = game.players.firstWhere((p) => p.isHuman).id;
       final bus = AppEventBus.create();
-      final appendFuture = awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(
-        bus,
-      );
+      final appendFuture =
+          awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(bus);
       autoConfirmDiplomacyDialogs(bus);
 
-      await pumpDiplomacyOrdersPanel(tester, game: game, humanId: humanId, bus: bus);
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: game,
+        humanId: humanId,
+        bus: bus,
+      );
 
       const actionLabels = <String>['Declare War', 'Offer Peace', 'Alliance'];
       Finder? actionFinder;
@@ -115,10 +96,17 @@ void main() {
       );
       final humanId = game.players.firstWhere((p) => p.isHuman).id;
       final bus = AppEventBus.create();
-      final breakFuture = awaitDiplomacyBusEvent<BreakAllianceImmediatelyEvent>(bus);
+      final breakFuture = awaitDiplomacyBusEvent<BreakAllianceImmediatelyEvent>(
+        bus,
+      );
       autoConfirmDiplomacyDialogs(bus);
 
-      await pumpDiplomacyOrdersPanel(tester, game: game, humanId: humanId, bus: bus);
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: game,
+        humanId: humanId,
+        bus: bus,
+      );
       await tapVisibleDiplomacy(tester, find.text('Break Alliance'));
 
       final event = await breakFuture;
@@ -134,9 +122,8 @@ void main() {
       final humanId = game.players.firstWhere((p) => p.isHuman).id;
       final target = game.players.firstWhere((p) => p.id != humanId).id;
       final bus = AppEventBus.create();
-      final removeFuture = awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(
-        bus,
-      );
+      final removeFuture =
+          awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(bus);
 
       await pumpDiplomacyOrdersPanel(
         tester,
@@ -160,73 +147,41 @@ void main() {
   );
 
   group('Boycott controls (Refs #3753 S14)', () {
-    testWidgets(
-      'colony holder shows Boycott enabled on GP row',
-      (WidgetTester tester) async {
-        final bus = AppEventBus.create();
-        final appendFuture =
-            awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(bus);
-        autoConfirmDiplomacyDialogs(bus);
+    testWidgets('colony holder shows Boycott enabled on GP row', (
+      WidgetTester tester,
+    ) async {
+      final bus = AppEventBus.create();
+      final appendFuture =
+          awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(bus);
+      autoConfirmDiplomacyDialogs(bus);
 
-        await pumpDiplomacyOrdersPanel(tester, game: _colonyGame(), bus: bus);
-        expect(find.text('Boycott'), findsOneWidget);
-        // Revoke Boycott is disabled until a boycott is active (Refs #4265 More).
-        expect(find.text('Revoke Boycott'), findsNothing);
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: diplomacyColonyGame(),
+        bus: bus,
+      );
+      expect(find.text('Boycott'), findsOneWidget);
+      // Revoke Boycott is disabled until a boycott is active (Refs #4265 More).
+      expect(find.text('Revoke Boycott'), findsNothing);
 
-        await tapVisibleDiplomacy(tester, find.text('Boycott'));
-        final event = await appendFuture;
-        expect(event.playerId, diplomacyOrdersHumanId);
-        expect(event.order.type, DiplomaticOrderType.boycott);
-        expect(event.order.targetFactionId, diplomacyOrdersGp2);
-      },
-    );
+      await tapVisibleDiplomacy(tester, find.text('Boycott'));
+      final event = await appendFuture;
+      expect(event.playerId, diplomacyOrdersHumanId);
+      expect(event.order.type, DiplomaticOrderType.boycott);
+      expect(event.order.targetFactionId, diplomacyOrdersGp2);
+    });
 
-    testWidgets(
-      'active boycott shows Revoke Boycott enabled on GP row',
-      (WidgetTester tester) async {
-        final bus = AppEventBus.create();
-        final appendFuture =
-            awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(bus);
-        autoConfirmDiplomacyDialogs(bus);
+    testWidgets('active boycott shows Revoke Boycott enabled on GP row', (
+      WidgetTester tester,
+    ) async {
+      final bus = AppEventBus.create();
+      final appendFuture =
+          awaitDiplomacyBusEvent<AppendDiplomaticOrderRequestedEvent>(bus);
+      autoConfirmDiplomacyDialogs(bus);
 
-        await pumpDiplomacyOrdersPanel(
-          tester,
-          game: _colonyGame(
-            boycotts: const [
-              BoycottState(
-                gpId: diplomacyOrdersHumanId,
-                targetGpId: diplomacyOrdersGp2,
-                sinceTurn: 1,
-              ),
-            ],
-          ),
-          bus: bus,
-        );
-        expect(find.text('Revoke Boycott'), findsOneWidget);
-
-        await tapVisibleDiplomacy(tester, find.text('Revoke Boycott'));
-        final event = await appendFuture;
-        expect(event.playerId, diplomacyOrdersHumanId);
-        expect(event.order.type, DiplomaticOrderType.revokeBoycott);
-        expect(event.order.targetFactionId, diplomacyOrdersGp2);
-      },
-    );
-
-    for (final c in <({
-      String name,
-      Game Function() game,
-      DiplomaticOrderType type,
-      String hiddenLabel,
-    })>[
-      (
-        name: 'pending boycott shows Cancel and removes on tap',
-        game: _colonyGame,
-        type: DiplomaticOrderType.boycott,
-        hiddenLabel: 'Boycott',
-      ),
-      (
-        name: 'pending revokeBoycott shows Cancel and removes on tap',
-        game: () => _colonyGame(
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: diplomacyColonyGame(
           boycotts: const [
             BoycottState(
               gpId: diplomacyOrdersHumanId,
@@ -235,10 +190,47 @@ void main() {
             ),
           ],
         ),
-        type: DiplomaticOrderType.revokeBoycott,
-        hiddenLabel: 'Revoke Boycott',
-      ),
-    ]) {
+        bus: bus,
+      );
+      expect(find.text('Revoke Boycott'), findsOneWidget);
+
+      await tapVisibleDiplomacy(tester, find.text('Revoke Boycott'));
+      final event = await appendFuture;
+      expect(event.playerId, diplomacyOrdersHumanId);
+      expect(event.order.type, DiplomaticOrderType.revokeBoycott);
+      expect(event.order.targetFactionId, diplomacyOrdersGp2);
+    });
+
+    for (final c
+        in <
+          ({
+            String name,
+            Game Function() game,
+            DiplomaticOrderType type,
+            String hiddenLabel,
+          })
+        >[
+          (
+            name: 'pending boycott shows Cancel and removes on tap',
+            game: diplomacyColonyGame,
+            type: DiplomaticOrderType.boycott,
+            hiddenLabel: 'Boycott',
+          ),
+          (
+            name: 'pending revokeBoycott shows Cancel and removes on tap',
+            game: () => diplomacyColonyGame(
+              boycotts: const [
+                BoycottState(
+                  gpId: diplomacyOrdersHumanId,
+                  targetGpId: diplomacyOrdersGp2,
+                  sinceTurn: 1,
+                ),
+              ],
+            ),
+            type: DiplomaticOrderType.revokeBoycott,
+            hiddenLabel: 'Revoke Boycott',
+          ),
+        ]) {
       testWidgets(c.name, (WidgetTester tester) async {
         final bus = AppEventBus.create();
         final removeFuture =
@@ -262,153 +254,147 @@ void main() {
       });
     }
 
-    testWidgets(
-      'Boycott disabled when human holds no colony',
-      (WidgetTester tester) async {
-        await pumpDiplomacyOrdersPanel(tester, game: buildDiplomacyPanelTestGame());
+    testWidgets('Boycott disabled when human holds no colony', (
+      WidgetTester tester,
+    ) async {
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: buildDiplomacyPanelTestGame(),
+      );
 
-        await tapVisibleDiplomacy(tester, find.text('More actions'));
-        final boycottButton = find.widgetWithText(CtNinePatchButton, 'Boycott');
-        expect(boycottButton, findsOneWidget);
-        expect(tester.widget<CtNinePatchButton>(boycottButton).enabled, isFalse);
-      },
-    );
+      await tapVisibleDiplomacy(tester, find.text('More actions'));
+      final boycottButton = find.widgetWithText(CtNinePatchButton, 'Boycott');
+      expect(boycottButton, findsOneWidget);
+      expect(tester.widget<CtNinePatchButton>(boycottButton).enabled, isFalse);
+    });
 
-    testWidgets(
-      'Minor row omits Boycott controls (negative)',
-      (WidgetTester tester) async {
-        await pumpDiplomacyOrdersPanel(
-          tester,
-          game: buildDiplomacyRichPanelTestGame(),
-          tall: true,
-          minorsTab: true,
-        );
+    testWidgets('Minor row omits Boycott controls (negative)', (
+      WidgetTester tester,
+    ) async {
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: buildDiplomacyRichPanelTestGame(),
+        tall: true,
+        minorsTab: true,
+      );
 
-        expect(find.text('Free City'), findsOneWidget);
-        expect(find.text('Boycott'), findsNothing);
-        expect(find.text('Revoke Boycott'), findsNothing);
-      },
-    );
+      expect(find.text('Free City'), findsOneWidget);
+      expect(find.text('Boycott'), findsNothing);
+      expect(find.text('Revoke Boycott'), findsNothing);
+    });
   });
 
   group('Minor subsidy / grant aid (Refs #3753 S15 / R2)', () {
-    testWidgets(
-      'pending setSubsidy shows percent line and Cancel',
-      (WidgetTester tester) async {
-        const percent = 15;
-        final bus = AppEventBus.create();
-        final removeFuture =
-            awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(bus);
+    testWidgets('pending setSubsidy shows percent line and Cancel', (
+      WidgetTester tester,
+    ) async {
+      const percent = 15;
+      final bus = AppEventBus.create();
+      final removeFuture =
+          awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(bus);
 
-        await pumpDiplomacyOrdersPanel(
-          tester,
-          game: buildDiplomacyRichPanelTestGame(),
-          bus: bus,
-          tall: true,
-          minorsTab: true,
-          currentOrders: diplomacyPendingOrders(
-            DiplomaticOrder(
-              type: DiplomaticOrderType.setSubsidy,
-              targetFactionId: diplomacyOrdersMinorId,
-              amount: percent,
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: buildDiplomacyRichPanelTestGame(),
+        bus: bus,
+        tall: true,
+        minorsTab: true,
+        currentOrders: diplomacyPendingOrders(
+          DiplomaticOrder(
+            type: DiplomaticOrderType.setSubsidy,
+            targetFactionId: diplomacyOrdersMinorId,
+            amount: percent,
+          ),
+        ),
+      );
+
+      expect(
+        find.text('Pending subsidy: $percent% (resolves end of turn)'),
+        findsOneWidget,
+      );
+      expect(find.text('Set Subsidy ($percent%)'), findsNothing);
+
+      await tapVisibleDiplomacy(tester, find.text('Cancel').first);
+      final event = await removeFuture;
+      expect(event.playerId, diplomacyOrdersHumanId);
+      expect(event.type, DiplomaticOrderType.setSubsidy);
+      expect(event.targetFactionId, diplomacyOrdersMinorId);
+    });
+
+    testWidgets('active subsidy shows outgoing percent line', (
+      WidgetTester tester,
+    ) async {
+      const percent = 10;
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: diplomacyMinorEmbassyGame(
+          subsidies: const [
+            SubsidyState(
+              payerId: diplomacyOrdersHumanId,
+              targetId: diplomacyOrdersMinorId,
+              percent: percent,
             ),
+          ],
+        ),
+        tall: true,
+        minorsTab: true,
+      );
+
+      expect(
+        find.text('Outgoing subsidy: $percent% to Free City'),
+        findsOneWidget,
+      );
+    });
+
+    testWidgets('pending grantAid shows amount line and Cancel', (
+      WidgetTester tester,
+    ) async {
+      const amount = 2000;
+      final bus = AppEventBus.create();
+      final removeFuture =
+          awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(bus);
+
+      await pumpDiplomacyOrdersPanel(
+        tester,
+        game: buildDiplomacyRichPanelTestGame(),
+        bus: bus,
+        tall: true,
+        minorsTab: true,
+        currentOrders: diplomacyPendingOrders(
+          DiplomaticOrder(
+            type: DiplomaticOrderType.grantAid,
+            targetFactionId: diplomacyOrdersMinorId,
+            amount: amount,
           ),
-        );
+        ),
+      );
 
-        expect(
-          find.text('Pending subsidy: $percent% (resolves end of turn)'),
-          findsOneWidget,
-        );
-        expect(find.text('Set Subsidy ($percent%)'), findsNothing);
-
-        await tapVisibleDiplomacy(tester, find.text('Cancel').first);
-        final event = await removeFuture;
-        expect(event.playerId, diplomacyOrdersHumanId);
-        expect(event.type, DiplomaticOrderType.setSubsidy);
-        expect(event.targetFactionId, diplomacyOrdersMinorId);
-      },
-    );
-
-    testWidgets(
-      'active subsidy shows outgoing percent line',
-      (WidgetTester tester) async {
-        const percent = 10;
-        await pumpDiplomacyOrdersPanel(
-          tester,
-          game: _minorEmbassyGame(
-            subsidies: const [
-              SubsidyState(
-                payerId: diplomacyOrdersHumanId,
-                targetId: diplomacyOrdersMinorId,
-                percent: percent,
-              ),
-            ],
+      final minorRow = diplomacyMinorRow();
+      expect(
+        find.descendant(
+          of: minorRow,
+          matching: find.text(
+            'Pending grant aid: £$amount (resolves end of turn)',
           ),
-          tall: true,
-          minorsTab: true,
-        );
+        ),
+        findsOneWidget,
+      );
+      expect(
+        find.descendant(of: minorRow, matching: find.text('Grant Aid')),
+        findsNothing,
+      );
 
-        expect(
-          find.text('Outgoing subsidy: $percent% to Free City'),
-          findsOneWidget,
-        );
-      },
-    );
+      await tapVisibleDiplomacy(
+        tester,
+        find.descendant(of: minorRow, matching: find.text('Cancel')),
+      );
+      final event = await removeFuture;
+      expect(event.playerId, diplomacyOrdersHumanId);
+      expect(event.type, DiplomaticOrderType.grantAid);
+      expect(event.targetFactionId, diplomacyOrdersMinorId);
+    });
 
-    testWidgets(
-      'pending grantAid shows amount line and Cancel',
-      (WidgetTester tester) async {
-        const amount = 2000;
-        final bus = AppEventBus.create();
-        final removeFuture =
-            awaitDiplomacyBusEvent<RemoveDiplomaticOrderRequestedEvent>(bus);
-
-        await pumpDiplomacyOrdersPanel(
-          tester,
-          game: buildDiplomacyRichPanelTestGame(),
-          bus: bus,
-          tall: true,
-          minorsTab: true,
-          currentOrders: diplomacyPendingOrders(
-            DiplomaticOrder(
-              type: DiplomaticOrderType.grantAid,
-              targetFactionId: diplomacyOrdersMinorId,
-              amount: amount,
-            ),
-          ),
-        );
-
-        final minorRow = diplomacyMinorRow();
-        expect(
-          find.descendant(
-            of: minorRow,
-            matching: find.text(
-              'Pending grant aid: £$amount (resolves end of turn)',
-            ),
-          ),
-          findsOneWidget,
-        );
-        expect(
-          find.descendant(of: minorRow, matching: find.text('Grant Aid')),
-          findsNothing,
-        );
-
-        await tapVisibleDiplomacy(
-          tester,
-          find.descendant(of: minorRow, matching: find.text('Cancel')),
-        );
-        final event = await removeFuture;
-        expect(event.playerId, diplomacyOrdersHumanId);
-        expect(event.type, DiplomaticOrderType.grantAid);
-        expect(event.targetFactionId, diplomacyOrdersMinorId);
-      },
-    );
-
-    for (final c in <({
-      String name,
-      String buttonLabel,
-      bool isSubsidy,
-    })>[
+    for (final c in <({String name, String buttonLabel, bool isSubsidy})>[
       (
         name: 'Grant Aid emits grantOrSubsidy dialog',
         buttonLabel: 'Grant Aid',
@@ -426,7 +412,7 @@ void main() {
 
         await pumpDiplomacyOrdersPanel(
           tester,
-          game: _minorEmbassyGame(),
+          game: diplomacyMinorEmbassyGame(),
           bus: bus,
           tall: true,
           minorsTab: true,
