+++ app/test/province_overlay_extraction_available_test.dart
@@ -4,7 +4,7 @@ import 'package:colonizethis_app_fixtures/demo/province_overlay_demo_data.dart'
 import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 import 'package:colonizethis_data/colonizethis_data.dart' show MapTopology;
 import 'package:colonizethis_logic/colonizethis_logic.dart'
-    show ProvinceImprovableCommodityCount, buildPlayerView;
+    show buildPlayerView;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/gestures.dart';
@@ -13,55 +13,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'province_overlay_test_harness.dart';
-
-String _foreignOwnedProvinceId({
-  required Game game,
-  required String humanPlayerId,
-}) {
-  for (final province in game.worldState.oldWorld.provinces) {
-    final ownerId = province.ownerId;
-    if (ownerId != null && ownerId.isNotEmpty && ownerId != humanPlayerId) {
-      return province.id;
-    }
-  }
-  fail(
-    'Test setup: demo game has no foreign-owned Old World province for '
-    'intel-gate pins.',
-  );
-}
-
-ProvinceExtractionSnapshot _sampleExtractionSnapshot(
-  String ownerId, {
-  int capitalGrainBonus = 0,
-}) {
-  return ProvinceExtractionSnapshot(
-    ownerId: ownerId,
-    capitalGrainBonus: capitalGrainBonus,
-    byCommodity: {
-      'grain': const ProvinceExtractionCommodityTotals(
-        effective: 1,
-        full: 5,
-        tileKeys: ['oldWorld|p1|0|0', 'oldWorld|p1|0|1'],
-      ),
-      'iron': const ProvinceExtractionCommodityTotals(
-        effective: 5,
-        full: 5,
-        tileKeys: ['oldWorld|p1|1|0'],
-      ),
-    },
-  );
-}
-
-const Map<String, ProvinceImprovableCommodityCount> _sampleAvailable = {
-  'grain': ProvinceImprovableCommodityCount(
-    count: 3,
-    tileKeys: ['oldWorld|p1|0|0', 'oldWorld|p1|2|0'],
-  ),
-  'timber': ProvinceImprovableCommodityCount(
-    count: 2,
-    tileKeys: ['oldWorld|p1|0|1'],
-  ),
-};
+import 'province_overlay_extraction_available_test_support.dart';
 
 void main() {
   suppressLogsForTests();
@@ -85,8 +37,8 @@ void main() {
         humanPlayerId: humanId,
         playerView: playerView,
         omniscientDetail: true,
-        extractionSnapshot: _sampleExtractionSnapshot(humanId),
-        availableByCommodity: _sampleAvailable,
+        extractionSnapshot: sampleProvinceExtractionSnapshot(humanId),
+        availableByCommodity: sampleProvinceImprovableAvailable,
       );
 
       expect(find.text('Extraction'), findsOneWidget);
@@ -124,8 +76,8 @@ void main() {
         humanPlayerId: humanId,
         playerView: playerView,
         omniscientDetail: true,
-        extractionSnapshot: _sampleExtractionSnapshot(humanId),
-        availableByCommodity: _sampleAvailable,
+        extractionSnapshot: sampleProvinceExtractionSnapshot(humanId),
+        availableByCommodity: sampleProvinceImprovableAvailable,
       );
 
       expect(
@@ -137,79 +89,71 @@ void main() {
     },
   );
 
-  testWidgets(
-    'full-yield Extraction omits partial reason line (Refs #4150)',
-    (tester) async {
-      final game = demoGameForOverlay;
-      final humanId = game.players.first.id;
-      final provinceId = ownedProvinceIdInOldWorld(
-        game: game,
+  testWidgets('full-yield Extraction omits partial reason line (Refs #4150)', (
+    tester,
+  ) async {
+    final game = demoGameForOverlay;
+    final humanId = game.players.first.id;
+    final provinceId = ownedProvinceIdInOldWorld(game: game, ownerId: humanId);
+    final playerView = buildPlayerView(game, const MapTopology(), humanId);
+
+    await pumpProvinceOverlayAtDarkTheme(
+      tester,
+      game: game,
+      displayId: provinceId,
+      region: demoRegionForOverlay,
+      humanPlayerId: humanId,
+      playerView: playerView,
+      omniscientDetail: true,
+      extractionSnapshot: ProvinceExtractionSnapshot(
         ownerId: humanId,
-      );
-      final playerView = buildPlayerView(game, const MapTopology(), humanId);
+        byCommodity: {
+          'grain': const ProvinceExtractionCommodityTotals(
+            effective: 5,
+            full: 5,
+            tileKeys: ['oldWorld|p1|0|0'],
+          ),
+        },
+      ),
+    );
 
-      await pumpProvinceOverlayAtDarkTheme(
-        tester,
-        game: game,
-        displayId: provinceId,
-        region: demoRegionForOverlay,
-        humanPlayerId: humanId,
-        playerView: playerView,
-        omniscientDetail: true,
-        extractionSnapshot: ProvinceExtractionSnapshot(
-          ownerId: humanId,
-          byCommodity: {
-            'grain': const ProvinceExtractionCommodityTotals(
-              effective: 5,
-              full: 5,
-              tileKeys: ['oldWorld|p1|0|0'],
-            ),
-          },
-        ),
-      );
+    expect(
+      find.text(
+        'Some improved tiles are not linked to your capital, or the road/port path is too weak.',
+      ),
+      findsNothing,
+    );
+  });
 
-      expect(
-        find.text(
-          'Some improved tiles are not linked to your capital, or the road/port path is too weak.',
-        ),
-        findsNothing,
-      );
-    },
-  );
+  testWidgets('empty Extraction shows dash placeholders (Refs #4002)', (
+    tester,
+  ) async {
+    final game = demoGameForOverlay;
+    final humanId = game.players.first.id;
+    final provinceId = ownedProvinceIdInOldWorld(game: game, ownerId: humanId);
+    final playerView = buildPlayerView(game, const MapTopology(), humanId);
 
-  testWidgets(
-    'empty Extraction shows dash placeholders (Refs #4002)',
-    (tester) async {
-      final game = demoGameForOverlay;
-      final humanId = game.players.first.id;
-      final provinceId = ownedProvinceIdInOldWorld(
-        game: game,
-        ownerId: humanId,
-      );
-      final playerView = buildPlayerView(game, const MapTopology(), humanId);
+    await pumpProvinceOverlayAtDarkTheme(
+      tester,
+      game: game,
+      displayId: provinceId,
+      region: demoRegionForOverlay,
+      humanPlayerId: humanId,
+      playerView: playerView,
+      omniscientDetail: true,
+    );
 
-      await pumpProvinceOverlayAtDarkTheme(
-        tester,
-        game: game,
-        displayId: provinceId,
-        region: demoRegionForOverlay,
-        humanPlayerId: humanId,
-        playerView: playerView,
-        omniscientDetail: true,
-      );
-
-      expect(find.text('Extraction'), findsOneWidget);
-      expect(find.text('Available'), findsOneWidget);
-      expect(find.text('—'), findsWidgets);
-    },
-  );
+    expect(find.text('Extraction'), findsOneWidget);
+    expect(find.text('Available'), findsOneWidget);
+    expect(find.text('—'), findsWidgets);
+  });
 
   testWidgets(
     'intel gate hides Extraction/Available quantities behind ??? (Refs #4002)',
     (tester) async {
       final game = demoGameForOverlay;
       final humanId = game.players.first.id;
-      final foreignProvinceId = _foreignOwnedProvinceId(
+      final foreignProvinceId = foreignOwnedProvinceIdForOverlay(
         game: game,
         humanPlayerId: humanId,
       );
@@ -223,8 +167,8 @@ void main() {
         humanPlayerId: humanId,
         playerView: playerView,
         omniscientDetail: false,
-        extractionSnapshot: _sampleExtractionSnapshot(humanId),
-        availableByCommodity: _sampleAvailable,
+        extractionSnapshot: sampleProvinceExtractionSnapshot(humanId),
+        availableByCommodity: sampleProvinceImprovableAvailable,
       );
 
       expect(find.text('Extraction'), findsNothing);
@@ -255,8 +199,8 @@ void main() {
         humanPlayerId: humanId,
         playerView: playerView,
         omniscientDetail: true,
-        extractionSnapshot: _sampleExtractionSnapshot(humanId),
-        availableByCommodity: _sampleAvailable,
+        extractionSnapshot: sampleProvinceExtractionSnapshot(humanId),
+        availableByCommodity: sampleProvinceImprovableAvailable,
         onHighlightTiles: (keys) {
           highlighted = keys;
         },
@@ -271,10 +215,7 @@ void main() {
       expect(mouseRegions, findsWidgets);
       final region = tester.widget<MouseRegion>(mouseRegions.first);
       region.onEnter!(const PointerEnterEvent());
-      expect(
-        highlighted,
-        ['oldWorld|p1|0|0', 'oldWorld|p1|0|1'],
-      );
+      expect(highlighted, ['oldWorld|p1|0|0', 'oldWorld|p1|0|1']);
       region.onExit!(const PointerExitEvent());
       expect(highlighted, isNull);
     },
@@ -353,11 +294,11 @@ void main() {
         humanPlayerId: humanId,
         playerView: playerView,
         omniscientDetail: true,
-        extractionSnapshot: _sampleExtractionSnapshot(
+        extractionSnapshot: sampleProvinceExtractionSnapshot(
           humanId,
           capitalGrainBonus: 2,
         ),
-        availableByCommodity: _sampleAvailable,
+        availableByCommodity: sampleProvinceImprovableAvailable,
       );
 
       expect(
@@ -371,57 +312,51 @@ void main() {
     },
   );
 
-  testWidgets(
-    'capital grain bonus annotation is not a hover-highlight target '
-    '(Refs #4064)',
-    (tester) async {
-      final game = demoGameForOverlay;
-      final humanId = game.players.first.id;
-      final provinceId = ownedProvinceIdInOldWorld(
-        game: game,
-        ownerId: humanId,
-      );
-      final playerView = buildPlayerView(game, const MapTopology(), humanId);
-      Iterable<String>? highlighted;
+  testWidgets('capital grain bonus annotation is not a hover-highlight target '
+      '(Refs #4064)', (tester) async {
+    final game = demoGameForOverlay;
+    final humanId = game.players.first.id;
+    final provinceId = ownedProvinceIdInOldWorld(game: game, ownerId: humanId);
+    final playerView = buildPlayerView(game, const MapTopology(), humanId);
+    Iterable<String>? highlighted;
 
-      await pumpProvinceOverlayAtDarkTheme(
-        tester,
-        game: game,
-        displayId: provinceId,
-        region: demoRegionForOverlay,
-        humanPlayerId: humanId,
-        playerView: playerView,
-        omniscientDetail: true,
-        extractionSnapshot: _sampleExtractionSnapshot(
-          humanId,
-          capitalGrainBonus: 2,
-        ),
-        availableByCommodity: _sampleAvailable,
-        onHighlightTiles: (keys) {
-          highlighted = keys;
-        },
-      );
+    await pumpProvinceOverlayAtDarkTheme(
+      tester,
+      game: game,
+      displayId: provinceId,
+      region: demoRegionForOverlay,
+      humanPlayerId: humanId,
+      playerView: playerView,
+      omniscientDetail: true,
+      extractionSnapshot: sampleProvinceExtractionSnapshot(
+        humanId,
+        capitalGrainBonus: 2,
+      ),
+      availableByCommodity: sampleProvinceImprovableAvailable,
+      onHighlightTiles: (keys) {
+        highlighted = keys;
+      },
+    );
 
-      final bonus = find.textContaining('incl. +2 capital grain bonus');
-      expect(bonus, findsOneWidget);
-      expect(
-        find.ancestor(of: bonus, matching: find.byType(MouseRegion)),
-        findsNothing,
-      );
+    final bonus = find.textContaining('incl. +2 capital grain bonus');
+    expect(bonus, findsOneWidget);
+    expect(
+      find.ancestor(of: bonus, matching: find.byType(MouseRegion)),
+      findsNothing,
+    );
 
-      final grainSegment = find.textContaining('1 (5)');
-      expect(grainSegment, findsOneWidget);
-      final grainRegions = find.ancestor(
-        of: grainSegment,
-        matching: find.byType(MouseRegion),
-      );
-      expect(grainRegions, findsWidgets);
-      tester.widget<MouseRegion>(grainRegions.first).onEnter!(
-        const PointerEnterEvent(),
-      );
-      expect(highlighted, ['oldWorld|p1|0|0', 'oldWorld|p1|0|1']);
-    },
-  );
+    final grainSegment = find.textContaining('1 (5)');
+    expect(grainSegment, findsOneWidget);
+    final grainRegions = find.ancestor(
+      of: grainSegment,
+      matching: find.byType(MouseRegion),
+    );
+    expect(grainRegions, findsWidgets);
+    tester.widget<MouseRegion>(grainRegions.first).onEnter!(
+      const PointerEnterEvent(),
+    );
+    expect(highlighted, ['oldWorld|p1|0|0', 'oldWorld|p1|0|1']);
+  });
 
   test(
     'setSecondaryHighlights stores multi keys and clears single (Refs #4002)',
