+++ tool/check_app_test_file_size.dart
@@ -19,25 +19,16 @@ const String _appTestsRelativePath = 'app/test';
 /// Remove an entry only after the file is at or under [_maxPhysicalLines].
 const List<String> appTestFileSizeAllowlistForTests = <String>[
   'app/test/app_event_handler_test.dart',
-  'app/test/debug_console_overlay_panel_test.dart',
   'app/test/diplomacy_panel_orders_test.dart',
-  'app/test/diplomacy_panel_test.dart',
   'app/test/game_event_bridge_test.dart',
-  'app/test/mobile_320dp_min_viewport_test.dart',
-  'app/test/move_army_invasion_intel_test.dart',
-  'app/test/move_dialogs_specs_part1_test.dart',
   'app/test/move_fleet_dialog_test.dart',
   'app/test/naval_panel_part1_pins.dart',
-  'app/test/panels_320dp_min_viewport_test.dart',
-  'app/test/pause_menu_side_menu_specs_test.dart',
-  'app/test/province_overlay_extraction_available_test.dart',
   'app/test/trade_screen_deal_book_tab_e6_test.dart',
   'app/test/trade_screen_market_tab_treasury_bid_cap_test.dart',
   'app/test/trade_screen_scaffold_test.dart',
   'app/test/units_panel_shared_widgets_test.dart',
   'app/test/units_panel_test_support_test.dart',
   'app/test/widgetbook_in_game_shell_chrome_test.dart',
-  'app/test/widgets/ct_nine_patch_button_dark_test.dart',
 ];
 
 int runCheckAppTestFileSize(
